Split MCP artifact export tests
